[#1087] Change the NTILE function to return Field<Integer> instead of Field<BigDecimal>

+++ jOOQ/src/main/java/org/jooq/impl/Factory.java
@@ -3750,8 +3750,8 @@ public class Factory implements FactoryOperations {
      * Sybase.
      */
     @Support({ DB2, POSTGRES, ORACLE, SQLSERVER, SYBASE })
-    public static WindowOverStep<BigDecimal> ntile(int number) {
-        return new WindowFunction<BigDecimal>("ntile", SQLDataType.NUMERIC, field("" + number, Integer.class));
+    public static WindowOverStep<Integer> ntile(int number) {
+        return new WindowFunction<Integer>("ntile", SQLDataType.INTEGER, field("" + number, Integer.class));
     }
 
     /**
